kafka_mock/mock-dspm-service/requirements.txt
kafka-python==2.0.2
pydantic==2.5.0
